+++ src/Functions/Utils/DateFormat.php
@@ -4,11 +4,11 @@ namespace FQL\Functions\Utils;
 
 use FQL\Functions;
 
-class DateFormat extends Functions\Core\MultipleFieldsFunction
+class DateFormat extends Functions\Core\SingleFieldFunction
 {
-    public function __construct(private readonly string $field, private readonly string $format = 'c')
+    public function __construct(string $field, private readonly string $format = 'c')
     {
-        parent::__construct($field, $format);
+        parent::__construct($field);
     }
 
     public function __invoke(array $item, array $resultItem): ?string