+++ src/Interface/Query.php
@@ -472,6 +472,7 @@ interface Query extends \Stringable
     public function cast(string $field, Type $as): Query;
     public function strToDate(string $valueField, string $format): Query;
     public function formatDate(string $dateField, string $format = 'c'): Query;
+    public function fromUnixTime(string $dateField, string $format = 'c'): Query;
     public function currentDate(bool $numeric = false): Query;
     public function currentTime(bool $numeric = false): Query;
     public function currentTimestamp(): Query;