+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Added code coverage report
 - [FiQuela CLI](https://github.com/1biot/fiquela-cli) is part of the main repository now
+- Added new function `FROM_UNIXTIME` for converting unix timestamp
 
 ### Changed
 - Remove support for php 8.1 and added support for php 8.5